update permission role for endpoint case & history
