+++ api/cases/route_prerequesites.js
@@ -1,83 +1,6 @@
 const replyHelper = require('../helpers')
 
 
-const CheckRoleView = server =>{
-    return {
-        method: (request, reply) =>{
-
-            if (   request.auth.credentials.user.role === "dinkesprov" 
-                || request.auth.credentials.user.role === "dinkeskota") {
-                return reply()
-            }else{
-                return reply({
-                    status: 403,
-                    message: 'Anda Tidak Mempunyai Akses!',
-                    data: null
-                }).code(403).takeover()
-            }
-           
-        },
-        assign: 'roles'
-    }
-}
-
-
-const CheckRoleCreate = server => {
-    return {
-        method: (request, reply) => {
-
-            if (request.auth.credentials.user.role === "dinkeskota") {
-                return reply()
-            } else {
-                return reply({
-                    status: 403,
-                    message: 'Anda Tidak Mempunyai Akses!',
-                    data: null
-                }).code(403).takeover()
-            }
-
-        },
-        assign: 'roles'
-    }
-}
-
-const CheckRoleUpdate = server => {
-    return {
-        method: (request, reply) => {
-
-            if (request.auth.credentials.user.role === "dinkeskota") {
-                return reply()
-            } else {
-                return reply({
-                    status: 403,
-                    message: 'Anda Tidak Mempunyai Akses!',
-                    data: null
-                }).code(403).takeover()
-            }
-
-        },
-        assign: 'roles'
-    }
-}
-
-const CheckRoleDelete = server => {
-    return {
-        method: (request, reply) => {
-            if (request.auth.credentials.user.role === "dinkeskota") {
-                return reply()
-            } else {
-                return reply({
-                    status: 403,
-                    message: 'Anda Tidak Mempunyai Akses!',
-                    data: null
-                }).code(403).takeover()
-            }
-
-        },
-        assign: 'roles'
-    }
-}
-
 const countCaseByDistrict = server =>{
     return {
         method: (request, reply) => {
@@ -110,10 +33,6 @@ const getCasebyId = server => {
 
 
 module.exports ={
-    CheckRoleView,
-    CheckRoleCreate,
-    CheckRoleUpdate,
-    CheckRoleDelete,
     countCaseByDistrict,
     getCasebyId
 }
