+++ api/histories/routes.js
@@ -3,10 +3,10 @@ module.exports = (server) =>{
     //const inputValidations = require('./validations/input')
     //const outputValidations = require('./validations/output') 
 
-    const CheckRoleView = require('../cases/route_prerequesites').CheckRoleView(server)
-    const CheckRoleCreate = require('../cases/route_prerequesites').CheckRoleCreate(server)
-    const CheckRoleUpdate = require('../cases/route_prerequesites').CheckRoleUpdate(server)
-
+   const CheckRoleView = require('../users/route_prerequesites').CheckRoleView(server)
+   const CheckRoleCreate = require('../users/route_prerequesites').CheckRoleCreate(server)
+   const CheckRoleUpdate = require('../users/route_prerequesites').CheckRoleUpdate(server)
+   const CheckRoleDelete = require('../users/route_prerequesites').CheckRoleDelete(server)
 
 
     return [
